# Conflicts:
#	OrderMicroservice/OrderService.Api/OrderService.Api.csproj
#	OrderMicroservice/OrderService.Application/Services/OrderService.cs

+++ OrderMicroservice/OrderService.Infrastructure/Repositories/OrderRepository.cs
@@ -68,7 +68,7 @@ public class OrderRepository : IOrderRepository
         return deletedOrder;
     }
 
-    public async Task AddItemToOrder(Guid orderId, int productId, string productName, decimal price, int quantity)
+    public async Task AddItemToOrder(Guid orderId, int productId, string productName, double price, int quantity)
     {
         var order = await GetByIdAsync(orderId);
